+++ src/calculator/polynomial_calculator.hpp
@@ -13,7 +13,7 @@ class PolynomialCalculator {
   PolynomialCalculator(Tokenizer& tokenizer, Parser& parser);
 
   Value process(std::string const& line);
-  void register_function(std::string const& name, size_t arity, std::function<Value(std::vector<Value> const&)> f);
+  void register_function(std::string const& name, unsigned long arity, std::function<Value(std::vector<Value> const&)> f);
   void register_constant(std::string const& name, Value value);
 
   Value last_value;
